UtilisateurDAO propre

+++ ProjetJava/src/DAO/UtilisateurDAO.java
@@ -26,41 +26,42 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
     @Override
     public Utilisateur create(Utilisateur obj, String motDePasse)
     {
+        obj = new Utilisateur();
         try
         {
+            if (obj.getNom().isEmpty())
+                throw new NullPointerException("ERROR : Nom vide");
+            if (obj.getPrenom().isEmpty())
+                throw new NullPointerException("ERROR : Prenom vide");
+            if (obj.getEmail().isEmpty())
+                throw new NullPointerException("ERROR : Email vide");
+            if (motDePasse.isEmpty())
+                throw new NullPointerException("ERROR : Mot de passe vide");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "INSERT INTO utilisateur (nom, prenom, email, mot_de_passe) "
                             + "VALUES( ? , ? , ? , ? )", Statement.RETURN_GENERATED_KEYS
                     );
-            if (obj.getNom().isEmpty())
-                prepare.setNull(1, Types.VARCHAR);
-            else
-                prepare.setString(1, obj.getNom());
 
-            if (obj.getPrenom().isEmpty())
-                prepare.setNull(2, Types.VARCHAR);
-            else
-                prepare.setString(2, obj.getPrenom());
-
-            if (obj.getEmail().isEmpty())
-                prepare.setNull(3, Types.VARCHAR);
-            else
-                prepare.setString(3, obj.getEmail());
-
+            prepare.setString(1, obj.getNom());
+            prepare.setString(2, obj.getPrenom());
+            prepare.setString(3, obj.getEmail());
             prepare.setString(4, motDePasse);
+
             prepare.executeUpdate();
 
             ResultSet result = prepare.getGeneratedKeys();
-            int id = 0;
-            if (result.next())
-                id = result.getInt(1);
+            if (!result.next())
+                throw new SQLException("SQL ERROR : ID autoIncrement nulle");
+
+            int id = result.getInt(1);
             obj = this.find(id);
 
             //}
-        } catch (SQLException e)
+        } catch (SQLException | NullPointerException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO create() " + e.getMessage());
         }
         return obj;
     }//end create
@@ -71,6 +72,9 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
         Utilisateur obj = new Utilisateur();
         try
         {
+            if (id == 0)
+                throw new NullPointerException("ERROR : ID NULLE");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "SELECT * FROM utilisateur WHERE id = ? ",
@@ -80,18 +84,19 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
             prepare.setInt(1, id);
             ResultSet result = prepare.executeQuery();
 
-            if (result.first())
-                obj = new Utilisateur(id,
-                        result.getString("nom"),
-                        result.getString("prenom"),
-                        result.getString("email"),
-                        result.getString("role")
-                );
-            
+            if (!result.next())
+                throw new SQLException("SQL ERROR : PAS DE RESULTAT");
 
-        } catch (SQLException e)
+            obj = new Utilisateur(id,
+                    result.getString("nom"),
+                    result.getString("prenom"),
+                    result.getString("email"),
+                    result.getString("role")
+            );
+
+        } catch (SQLException | NullPointerException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO find() " + e.getMessage());
         }
         return obj;
     }//end find
@@ -101,6 +106,15 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
     {
         try
         {
+            if (obj.getId() == 0)
+                throw new NullPointerException("ERROR : ID nulle");
+            if (obj.getNom().isEmpty())
+                throw new NullPointerException("ERROR : Nom vide");
+            if (obj.getPrenom().isEmpty())
+                throw new NullPointerException("ERROR : Prenom vide");
+            if (obj.getEmail().isEmpty())
+                throw new NullPointerException("ERROR : Mail vide");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "UPDATE utilisateur "
@@ -109,41 +123,35 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
                             ResultSet.TYPE_SCROLL_INSENSITIVE,
                             ResultSet.CONCUR_UPDATABLE
                     );
-            if (obj.getNom().isEmpty())
-                prepare.setNull(1, Types.VARCHAR);
-            else
-                prepare.setString(1, obj.getNom());
 
-            if (obj.getPrenom().isEmpty())
-                prepare.setNull(2, Types.VARCHAR);
-            else
-                prepare.setString(2, obj.getPrenom());
-
-            if (obj.getEmail().isEmpty())
-                prepare.setNull(3, Types.VARCHAR);
-            else
-                prepare.setString(3, obj.getEmail());
-
+            prepare.setString(1, obj.getNom());
+            prepare.setString(2, obj.getPrenom());
+            prepare.setString(3, obj.getEmail());
             prepare.setInt(4, obj.getId());
 
             prepare.executeUpdate();
 
             obj = this.find(obj.getId());
-        } catch (SQLException e)
+        } catch (SQLException | NullPointerException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO update() " + e.getMessage());
         }
 
         return obj.getId() != 0;
     }//end update
-    
+
     public boolean modifierMotDePasse(Utilisateur obj, String ancienMotDePasse, String nouveauMotDePasse)
     {
-        if(connexion(obj.getEmail(), ancienMotDePasse).getId() == 0)
+        if (connexion(obj.getEmail(), ancienMotDePasse).getId() == 0)
             return false;
-        
+
         try
         {
+            if (obj.getId() == 0)
+                throw new NullPointerException("ERROR : ID nulle");
+            if (nouveauMotDePasse.isEmpty())
+                throw new NullPointerException("ERROR : Nouveau mot de passe vide");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "UPDATE utilisateur "
@@ -152,29 +160,20 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
                             ResultSet.TYPE_SCROLL_INSENSITIVE,
                             ResultSet.CONCUR_UPDATABLE
                     );
-          
-            if (nouveauMotDePasse.isEmpty())
-                prepare.setNull(1, Types.CHAR);
-            else
-            {                
-                try
-                {
-                    MessageDigest digest = MessageDigest.getInstance("SHA-256");
-                    byte[] hash = digest.digest(nouveauMotDePasse.getBytes(StandardCharsets.UTF_8));
-                    nouveauMotDePasse = new String(Hex.encode(hash));
-                } catch (NoSuchAlgorithmException ex)
-                {
-                }
-                prepare.setString(1, nouveauMotDePasse);
-            }
-            
+
+            MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            byte[] hash = digest.digest(nouveauMotDePasse.getBytes(StandardCharsets.UTF_8));
+            nouveauMotDePasse = new String(Hex.encode(hash));
+
+            prepare.setString(1, nouveauMotDePasse);
             prepare.setInt(2, obj.getId());
 
             prepare.executeUpdate();
 
-        } catch (SQLException e)
+        } catch (SQLException | NullPointerException | NoSuchAlgorithmException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO modifierMotDePasse() " + e.getMessage());
+            return false;
         }
 
         return true;
@@ -185,6 +184,9 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
     {
         try
         {
+            if (obj.getId() == 0)
+                throw new NullPointerException("ERROR : ID nulle");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "DELETE FROM utilisateur WHERE id = ? ",
@@ -194,9 +196,9 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
             prepare.setInt(1, obj.getId());
             prepare.executeUpdate();
 
-        } catch (SQLException e)
+        } catch (SQLException | NullPointerException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO delete() " + e.getMessage());
         }
     }//end delete
 
@@ -205,72 +207,63 @@ public class UtilisateurDAO extends DAO<Utilisateur, String>
         Utilisateur obj = new Utilisateur();
         try
         {
+            if (email.isEmpty())
+                throw new NullPointerException("ERROR : Email vide");
+            if (motDePasse.isEmpty())
+                throw new NullPointerException("ERROR : Mot de passe vide");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "SELECT * FROM utilisateur WHERE email = ? AND mot_de_passe = ? ",
                             ResultSet.TYPE_SCROLL_INSENSITIVE,
                             ResultSet.CONCUR_UPDATABLE
                     );
-            if (email.isEmpty())
-                prepare.setNull(1, Types.VARCHAR);
-            else
-                prepare.setString(1, email);
 
-            if (motDePasse.isEmpty())
-                prepare.setNull(2, Types.CHAR);
-            else
-            {                
-                try
-                {
-                    MessageDigest digest = MessageDigest.getInstance("SHA-256");
-                    byte[] hash = digest.digest(motDePasse.getBytes(StandardCharsets.UTF_8));
-                    motDePasse = new String(Hex.encode(hash));
-                } catch (NoSuchAlgorithmException ex)
-                {
-                }
-                prepare.setString(2, motDePasse);
-            }
+            prepare.setString(1, email);
+
+            MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            byte[] hash = digest.digest(motDePasse.getBytes(StandardCharsets.UTF_8));
+            motDePasse = new String(Hex.encode(hash));
+
+            prepare.setString(2, motDePasse);
 
             ResultSet result = prepare.executeQuery();
 
-            int id = 0; // 
-            
             if (result.next())
-                id = result.getInt(1);
-            
-            obj = this.find(id);
+                obj = this.find(result.getInt(1));
 
-        } catch (SQLException e)
+        } catch (SQLException | NullPointerException | NoSuchAlgorithmException e)
         {
-            e.printStackTrace();
+            System.err.println("UtilisateurDAO connexion() " + e.getMessage());
         }
         return obj;
     }
-    
+
     public boolean mailExiste(String email)
     {
         try
         {
+            if (email.isEmpty())
+                throw new NullPointerException("ERROR : Email vide");
+
             PreparedStatement prepare = this.connect
                     .prepareStatement(
                             "SELECT * FROM utilisateur WHERE email = ? ",
                             ResultSet.TYPE_SCROLL_INSENSITIVE,
                             ResultSet.CONCUR_UPDATABLE
                     );
-            if (email.isEmpty())
-                prepare.setNull(1, Types.VARCHAR);
-            else
-                prepare.setString(1, email);
-            
+
+            prepare.setString(1, email);
+
             ResultSet result = prepare.executeQuery();
-            
-            return result.first();//true / false
-            
-        } catch (SQLException ex)
+
+            return result.next();//true / false
+
+        } catch (SQLException | NullPointerException e)
         {
-            
+            System.err.println("UtilisateurDAO mailExiste() " + e.getMessage());
         }
-        
+
         return true;
     }
 }//end classe
